qc/pipeline: move the 'copy_project' function into 'analysis' module (& rename to 'copy_analysis_project').
